Update security configuration and Keycloak integration
The security configuration has been updated to enable method security. Permissions are now handled using the @PreAuthorize annotation in the controller instead of the security configuration class. Furthermore, Keycloak integration has been adjusted with a realm import and the addition of a custom role (ROLE_STAFF) in the Keycloak realm configuration.